Updated to detect whether or not the data file includes versions.

+++ main.py
@@ -29,25 +29,27 @@ def create_skill_lines(skills: List[str]) -> List[str]:
     return out
 
 
-# Format Technical Skill Lines
-stack_lines = create_skill_lines(data['technical_skills']['solution_stack'])
-software_lines = create_skill_lines(data['technical_skills']['software_tools'])
+# Handle Implicit Single Version Case
+if 'versions' not in data:
+    # Format Technical Skill Lines
+    stack_lines = create_skill_lines(data['technical_skills']['solution_stack'])
+    software_lines = create_skill_lines(data['technical_skills']['software_tools'])
 
-# Fill out jinja template
-environment = Environment(loader=FileSystemLoader('templates/'), trim_blocks=True, lstrip_blocks=True)
-template = environment.get_template('resume.txt')
-content = template.render(
-    data,
-    stack_lines=stack_lines,
-    software_lines=software_lines,
-)
+    # Fill out jinja template
+    environment = Environment(loader=FileSystemLoader('templates/'), trim_blocks=True, lstrip_blocks=True)
+    template = environment.get_template('resume.txt')
+    content = template.render(
+        data,
+        stack_lines=stack_lines,
+        software_lines=software_lines,
+    )
 
-# Write output to file
-with open('output/raw_tex/resume.tex', 'w') as outfile:
-    outfile.write(content)
+    # Write output to file
+    with open('output/raw_tex/resume.tex', 'w') as outfile:
+        outfile.write(content)
 
-# Compile latex file into appropriate folder
-os.system("pdflatex -output-directory output\\compiled_results output\\raw_tex\\resume.tex")
+    # Compile latex file into appropriate folder
+    os.system("pdflatex -output-directory output\\compiled_results output\\raw_tex\\resume.tex")
 
-# Remove logfile
-os.remove('output\\compiled_results\\resume.log')
+    # Remove logfile
+    os.remove('output\\compiled_results\\resume.log')
